requirements.txt
asgiref==3.8.1
Django==5.1.1
django-allauth==65.0.2
django-querycount==0.8.3
djangorestframework==3.15.2
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
